movieWatchDjango/requirements.txt
attrs==19.3.0
backcall==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.shutil-which==3.5.2
bleach==3.1.4
boto3==1.12.44
botocore==1.15.44
cython
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
configparser==4.0.2
contextlib2==0.6.0.post1
decorator==4.4.2
defusedxml==0.6.0
Django==2.2
django-cors-headers==3.0.2
django-extensions==3.1.1
django-rest-framework==0.1.0
django-rest-multiple-models==2.1.3
djangorestframework==3.9.4
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
enum34==1.1.10
futures
gunicorn==19.9.0
httpie==2.0.0
idna==2.9
importlib-metadata==1.6.0
ipaddress==1.0.23
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
mysql-connector==2.2.9
nbconvert==5.6.1
nbformat==4.4.0
notebook==5.7.9
pandocfilters==1.4.2
parso==0.7.0
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.10
prometheus-client==0.7.1
prompt-toolkit==1.0.18
ptyprocess==0.6.0
Pygments==2.5.2
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
qtconsole==4.7.3
QtPy==1.9.0
requests==2.23.0
s3transfer==0.3.3
scandir==1.10.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
spotipy
terminado==0.8.3
testpath==0.4.4
tmdb3==0.7.2
tmdbsimple==2.2.0
tornado==5.1.1
traitlets==4.3.3
typing==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
yarg==0.1.9
zipp==1.2.0
